refactor: moved SpeechBubbleShape.kt to theme package
